Better encapsulation
I'm going to break out some of the functions in automatebeta once I've
eliminated threadpool.
