+++ Stitch.py
@@ -1,4 +1,4 @@
-import os, Image, ImageFile, sys, threading
+import os, Image, ImageFile, sys, threading, threadpool, numpy
 try:
 	import psyco
 	psyco.full()
@@ -8,8 +8,13 @@ except:
 targetDimension = 280
 targetSize = 512 * 1024 # 512 KB
 
-from autocrop import autocrop
-import threadpool
+_imageDtype = numpy.dtype('i')
+def autocrop(img):
+	alpha = numpy.array(img, dtype=_imageDtype)[:,:,3]
+	horizontal = alpha.any(axis=0).nonzero()[0]
+	vertical = alpha.any(axis=1).nonzero()[0]
+	cropping = (horizontal[0], vertical[0], horizontal[-1], vertical[-1])
+	return (img.crop(cropping), cropping)
 
 def cropTask(i, s, filename):
 	# print 'Processing:', filename
